Move detector.
