+++ crates/core/src/detector.rs
@@ -1,6 +1,11 @@
+use crate::color;
+use crate::config::{Config, CONFIG_NAME};
 use crate::errors::ProtoError;
+use crate::manifest::{Manifest, MANIFEST_NAME};
+use crate::tool::Tool;
 use lenient_semver::Version;
-use std::{fs, path::Path};
+use log::{debug, trace};
+use std::{env, fs, path::Path};
 
 #[async_trait::async_trait]
 pub trait Detector<'tool>: Send + Sync {
@@ -15,6 +20,127 @@ pub fn load_version_file(path: &Path) -> Result<String, ProtoError> {
         .map_err(|e| ProtoError::Fs(path.to_path_buf(), e.to_string()))?
         .trim()
         .to_owned())
+}
+
+pub async fn detect_version_from_environment(
+    tool: &Box<dyn Tool<'_>>,
+    forced_version: Option<String>,
+) -> Result<String, ProtoError> {
+    let mut version = forced_version;
+    let env_var = format!("PROTO_{}_VERSION", tool.get_bin_name().to_uppercase());
+
+    // Env var takes highest priority
+    if version.is_none() {
+        if let Ok(session_version) = env::var(&env_var) {
+            debug!(
+                target: "proto:detector",
+                "Detected version {} from environment variable {}",
+                session_version,
+                env_var
+            );
+
+            version = Some(session_version);
+        }
+    } else {
+        debug!(
+            target: "proto:detector",
+            "Using explicit version {} passed on the command line",
+            version.as_ref().unwrap(),
+        );
+    }
+
+    // Traverse upwards and attempt to detect a local version
+    if let Ok(working_dir) = env::current_dir() {
+        trace!(
+            target: "proto:detector",
+            "Attempting to find local version"
+        );
+
+        let mut current_dir: Option<&Path> = Some(&working_dir);
+
+        while let Some(dir) = &current_dir {
+            trace!(
+                target: "proto:detector",
+                "Checking in directory {}",
+                color::path(dir)
+            );
+
+            // We already found a version, so exit
+            if version.is_some() {
+                break;
+            }
+
+            // Detect from our config file
+            trace!(
+                target: "proto:detector",
+                "Checking proto configuration file ({})",
+                CONFIG_NAME
+            );
+
+            let config = Config::load_from(dir)?;
+
+            if let Some(local_version) = config.tools.get(tool.get_bin_name()) {
+                debug!(
+                    target: "proto:detector",
+                    "Detected version {} from configuration file {}",
+                    local_version,
+                    color::path(&config.path)
+                );
+
+                version = Some(local_version.to_owned());
+                break;
+            }
+
+            // Detect using the tool
+            trace!(
+                target: "proto:detector",
+                "Detecting from the tool's ecosystem"
+            );
+
+            if let Some(eco_version) = tool.detect_version_from(dir).await? {
+                debug!(
+                    target: "proto:detector",
+                    "Detected version {} from tool's ecosystem",
+                    eco_version,
+                );
+
+                version = Some(eco_version);
+                break;
+            }
+
+            current_dir = dir.parent();
+        }
+    }
+
+    // If still no version, load the global version
+    if version.is_none() {
+        trace!(
+            target: "proto:detector",
+            "Attempting to find global version in manifest ({})",
+            MANIFEST_NAME
+        );
+
+        let manifest = Manifest::load_for_tool(tool.get_bin_name())?;
+
+        if let Some(global_version) = manifest.default_version {
+            debug!(
+                target: "proto:detector",
+                "Detected global version {} from {}",
+               global_version,
+                color::path(&manifest.path)
+            );
+
+            version = Some(global_version);
+        }
+    }
+
+    // We didn't find anything!
+    match version {
+        Some(ver) => Ok(ver),
+        None => Err(ProtoError::Message(
+            "Unable to detect an applicable version. Try setting a local or global version, or passing a command line argument.".into(),
+        )),
+    }
 }
 
 pub fn get_fixed_version(version: &str) -> Option<String> {
